+++ cxroots/tests/test_rootfinding_contours.py
@@ -2,7 +2,7 @@ import unittest
 from numpy import pi
 
 from cxroots import Circle, Rectangle, AnnulusSector, Annulus
-from cxroots.tests.ApproxEqual import roots_approx_equal
+from cxroots.tests.approx_equal import roots_approx_equal
 
 
 class TestRootfindingContours(unittest.TestCase):
